Add files via upload
A correct prediction of stocks can lead to huge profits for the seller and the broker.
Frequently, it is brought out that prediction is chaotic rather than random, which means it
can be predicted by carefully analyzing the history of the respective stock market. Machine
learning is an efficient way to represent such processes. It predicts a market value close to
the tangible value, thereby increasing the accuracy. The introduction of machine learning to the
area of stock prediction has appealed to many researchers because of its efficient and
accurate measurements [1] [2].